backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
redis>=5.0.1
httpx>=0.26.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.3
aiofiles>=23.2.1
qdrant-client>=1.7.0
numpy>=1.26.0
jinaai>=0.0.42
google-generativeai>=0.3.1
psycopg2-binary>=2.9.9
websockets>=12.0
uuid>=1.30
requests==2.31.0